Pulls out FindOutlierThreshold into own service
