+++ src/firefly_aws/domain/service/check_resource_usage.py
@@ -17,6 +17,7 @@ class CheckResourceUsage(ff.DomainService):
     _message_transport: ff.MessageTransport = None
     _configuration: ff.Configuration = None
     _context: str = None
+    _find_outlier_threshold = domain.FindOutlierThreshold = None
 
     def __init__(self):
         context = self._configuration.contexts[self._context]
@@ -69,37 +70,3 @@ class CheckResourceUsage(ff.DomainService):
             lower_max = self._memory_settings[memory_index]
             if outlier_threshold < (lower_max * .9): # If threshold is less than 90% of lower_max, then safe to drop down memory tier
                 self._resource_monitor.set_memory_level(str(message), self._memory_settings[memory_index])
-    
-    def _find_outlier_threshold(self, memory_list, deviation):
-        sorted_memory_usage = sorted(memory_list)
-        median_memory_usage = median(sorted_memory_usage)
-
-        absolute_deviation_from_median = [abs(value - median_memory_usage) for value in sorted_memory_usage]
-        
-        median_absolute_deviation = median(absolute_deviation_from_median)
-
-        outlier_threshold = 0
-        
-        # If more than 50% of our values are the same value, MAD will be 0. Then initiate backup plan with z_score (very rare edgecase)
-        # Not the best case to use z-score, but good backup
-        if median_absolute_deviation == 0:
-            outlier_threshold = self._find_z_score_threshold(memory_list, deviation)
-        else:
-            # 1.4826 is a constant linked to assumption of normally distributed data excluding outliers
-            memory_mad = median_absolute_deviation * 1.4826
-
-            for value in sorted_memory_usage:
-                memory_mad_relative_deviation = (value - median_memory_usage) / memory_mad
-                # Any relative deviation greater than passed in deviation is an outlier
-                if abs(memory_mad_relative_deviation) < deviation:
-                    #since we're iterating through a sorted list, the value can always be updated if true
-                    outlier_threshold = value
-
-        return outlier_threshold
-
-    def _find_z_score_threshold(self, memory_list, deviation):
-        standard_deviation = stdev(memory_list)
-        average = sum(memory_list) / len(memory_list)
-        outlier_threshold = (standard_deviation * deviation) + average
-
-        return outlier_threshold
